% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{admins_code_validate}
\alias{admins_code_validate}
\title{Administration code varidation}
\usage{
admins_code_validate(jis_code)
}
\arguments{
\item{jis_code}{jis code for prefecture and city identifical number.
If prefecture, must be from 1 to 47. If city, range of 5 digits.}
}
\description{
Administration code varidation
}
