% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_validate.R
\name{code_reform}
\alias{code_reform}
\title{Reform input jis code as 2 or 5 character length.}
\usage{
code_reform(jis_code)
}
\arguments{
\item{jis_code}{jis code for prefecture and city identifical number.
If prefecture, must be from 1 to 47. If city, range of 5 digits.}
}
\description{
Reform input jis code as 2 or 5 character length.
}
\note{
The \code{code_reform} function was added in version 0.3.2.9000
}
\examples{
code_reform(c(1, "33", "08201"))
}
