% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_validate.R
\name{code_validate}
\alias{code_validate}
\title{Administration code varidation}
\usage{
code_validate(jis_code)
}
\arguments{
\item{jis_code}{jis code for prefecture and city identifical number.
If prefecture, must be from 1 to 47. If city, range of 5 digits.}
}
\description{
Administration code varidation
}
\note{
The \code{code_validate} function was added in version 0.3.2.9000
}
\examples{
\dontrun{
code_validate(jis_code = "05")
code_validate(jis_code = 33101)
code_validate(jis_code = c("01", "33101"))
}
}
