% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{collect_cityarea}
\alias{collect_cityarea}
\title{Collect administration area}
\usage{
collect_cityarea(path = NULL)
}
\arguments{
\item{path}{path to N03 shapefile (if already exist)}
}
\description{
Collect administration area
}

