% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{collect_ksj_p33}
\alias{collect_ksj_p33}
\title{Collect administration office point datasets.}
\usage{
collect_ksj_p33(code = NULL, path = NULL)
}
\arguments{
\item{code}{prefecture code}

\item{path}{shepfile directory}
}
\description{
Collect administration office point datasets.
}

