% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{collect_ksj_p34}
\alias{collect_ksj_p34}
\title{Collect administration office point datasets.}
\usage{
collect_ksj_p34(path = NULL)
}
\arguments{
\item{path}{path to N03 shapefile (if already exist)}
}
\description{
Collect administration office point datasets.
}

