% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{collect_prefcode}
\alias{collect_prefcode}
\title{Get prefecture code (JIS X 0402)}
\usage{
collect_prefcode(code = NULL, admin_name = NULL)
}
\arguments{
\item{code}{numeric}

\item{admin_name}{prefecture code for Japanese (character)}
}
\description{
Get prefecture code from prefecture of name or number.
}

