% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_viewer.R
\name{district_viewer}
\alias{district_viewer}
\title{District Viewer}
\usage{
district_viewer(code = 33, color = "red")
}
\arguments{
\item{code}{prefecture code (default 33)}

\item{color}{polygon line color for leaflet}
}
\description{
Interactive district map and information tool.
}
\examples{
\dontrun{
district_viewer()
}
}
