% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city.R
\name{find_city}
\alias{find_city}
\title{Detect city by coordinates}
\usage{
find_city(lon = lon, lat = lat, ...)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{...}{path to arguments \code{spdf_jpn_pref}.}
}
\description{
Detect city by coordinates
}
\note{
The \code{find_city} function was added in version 0.2.1
}
