% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_name_table.R
\name{find_jis_code}
\alias{find_jis_code}
\title{Find jis city code}
\usage{
find_jis_code(pref_code, admin_name, strict = TRUE)
}
\arguments{
\item{pref_code}{jis code from 1 to 47}

\item{admin_name}{prefecture names (string)}

\item{strict}{matching patterns}
}
\value{
Identification code for cities,towns and villages (JIS X 0402:2010)
}
\description{
Find jis city code
}
\examples{
\dontrun{
find_jis_code(33, "\\u5009\\u6577\\u5e02")
find_jis_code(33, "\\u5009\\u6577\\u5e02", strict = FALSE)

find_jis_code(14, c("\\u938c\\u5009\\u5e02", "\\u5c0f\\u7530\\u539f\\u5e02"), strict = FALSE)
}
}
