% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city.R
\name{find_pref}
\alias{find_pref}
\title{Detect prefecture by coordinates}
\usage{
find_pref(longitude, latitude)
}
\arguments{
\item{longitude}{longitude}

\item{latitude}{latitude}
}
\description{
Detect prefecture by coordinates
}
\note{
The \code{find_pref} function was added in version 0.2.1
}
\examples{
\dontrun{
find_pref(longitude = 122.940625, latitude = 24.4520833334)
}
}
