% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city.R
\name{find_prefs}
\alias{find_prefs}
\title{Detect prefectures by coordinates}
\usage{
find_prefs(lon = lon, lat = lat)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}
}
\description{
Detect prefectures by coordinates
}
