% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jpn.R
\name{gg_jpn_district}
\alias{gg_jpn_district}
\title{Display city map based on 'ggplot2'}
\usage{
gg_jpn_district(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Rendering city map using 'ggplot2' and 'ggghost' system.
}
\examples{
\dontrun{
gg_jpn_district(spdf_jpn_pref(code = "37", district = TRUE))
}
}

