% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdf_jpn.R
\name{jpn_admins}
\alias{jpn_admins}
\title{Simple features for administration office points}
\usage{
jpn_admins(path = NULL, jis_code)
}
\arguments{
\item{path}{shapefile path}

\item{jis_code}{jis code for prefecture and city identifical number}
}
\value{
data.frame. contains follow columns jis_code, type, name, address, longitude and latitude.
}
\description{
Name and geolocations for administration offices in prefecture.
}
\examples{
\dontrun{
jpn_admins(jis_code = 17)
}
}
