% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jpnprefs}
\alias{jpnprefs}
\title{Prefectural informations in Japan}
\format{A data frame with X rows Y variables:
\itemize{
  \item{prefecture: prefecture names}
  \item{capital: capital name for prefecture}
  \item{region: region}
  \item{jis_code: jis code}
  \item{major_island: }
}}
\usage{
jpnprefs
}
\description{
Prefectural informations in Japan
}
\keyword{datasets}

