% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{path_ksj_cityarea}
\alias{path_ksj_cityarea}
\title{Download KSJ N03 zip files}
\usage{
path_ksj_cityarea(code = NULL, path = NULL)
}
\arguments{
\item{code}{prefecture code (JIS X 0402)}

\item{path}{path to N03 shapefile (if already exist)}
}
\description{
Download KSJ N03 zip files
}
