% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prefecture_mesh}
\alias{prefecture_mesh}
\title{Prefecture's meshcode}
\format{A simple feature data frame with 314 rows 6 variables:
\itemize{
\item{pref_code: prefecture code}
\item{prefecture: name}
\item{city_code: city code (JIS code)}
\item{city: name}
\item{geometry}
}}
\usage{
prefecture_mesh
}
\description{
Prefectures dataset.
}
\keyword{datasets}
