% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_ksj_cityarea}
\alias{read_ksj_cityarea}
\title{Intermediate function}
\usage{
read_ksj_cityarea(code = NULL, path = NULL)
}
\arguments{
\item{code}{prefecture code (JIS X 0402)}

\item{path}{path to N03 shapefile (if already exist)}
}
\description{
Download N03 raw data files or loading if file exists.
}
