% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksj.R
\name{read_ksj_p33}
\alias{read_ksj_p33}
\title{Read administration office point datasets.}
\usage{
read_ksj_p33(code = NULL, path = NULL)
}
\arguments{
\item{code}{prefecture code}

\item{path}{shapefile path}
}
\description{
Name and geolocations for administration offices in prefecture.
}
\examples{
\dontrun{
read_ksj_p33(code = 17)
}
}

