% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_ksj_p34}
\alias{read_ksj_p34}
\title{Intermediate function}
\usage{
read_ksj_p34(pref_code = NULL, path = NULL)
}
\arguments{
\item{pref_code}{prefecture code (JIS X 0402)}

\item{path}{path to P34 shapefile (if already exist)}
}
\description{
Intermediate function
}
