% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdf_jpn.R
\name{spdf_jpn_admins}
\alias{spdf_jpn_admins}
\title{Spatial Dataframe for administration office points}
\usage{
spdf_jpn_admins(path, code = NULL, jis_code_city = NULL)
}
\arguments{
\item{path}{shapefile path}

\item{code}{prefecture code}

\item{jis_code_city}{jis code for city as jis_code_pref + identifier number}
}
\value{
data.frame. contains follow columns jis_code, type, name, address, longitude and latitude.
}
\description{
Name and geolocations for administration offices in prefecture.
}
\examples{
\dontrun{
spdf_jpn_admins(code = 17)
}
}

