% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdf_jpn.R
\name{spdf_jpn_cities}
\alias{spdf_jpn_cities}
\title{Spatial Data frame cities}
\usage{
spdf_jpn_cities(jis_code_pref, jis_code = NULL, admin_name = NULL)
}
\arguments{
\item{jis_code_pref}{numeric}

\item{jis_code}{jis code}

\item{admin_name}{string}
}
\description{
Spatial Data frame cities
}
\examples{
\dontrun{
spdf_jpn_cities(jis_code_pref = 33, jis_code = 33103)
spdf_jpn_cities(jis_code_pref = 33, jis_code = c(33103, 33104, 33205))
}
}

