% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdf_jpn.R
\name{spdf_jpn_pref}
\alias{spdf_jpn_pref}
\title{Spatial Data frame}
\usage{
spdf_jpn_pref(code = NULL, admin_name = NULL, district = TRUE)
}
\arguments{
\item{code}{jis code}

\item{admin_name}{string}

\item{district}{logical (default TRUE)}
}
\description{
Spatial Data frame
}
\examples{
\dontrun{
spdf_jpn_pref(admin_name = "岡山県", district = FALSE)
spdf_jpn_pref(code = 33, district = FALSE)
}
}

