% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city.R
\name{which_pol_min}
\alias{which_pol_min}
\title{Internal function}
\usage{
which_pol_min(longitude, latitude)
}
\arguments{
\item{longitude}{longitude}

\item{latitude}{latitude}
}
\description{
Internal function
}
