% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bind_cityareas}
\alias{bind_cityareas}
\title{Bind city area polygons to prefecture polygon}
\usage{
bind_cityareas(path = NULL)
}
\arguments{
\item{path}{path to N03 shapefile (if already exist)}
}
\description{
Bind city area polygons to prefecture polygon
}
