% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city.R
\name{find_city}
\alias{find_city}
\title{Detect city by coordinates}
\usage{
find_city(longitude, latitude, ...)
}
\arguments{
\item{longitude}{longitude}

\item{latitude}{latitude}

\item{...}{path to arguments \code{spdf_jpn_pref}.}
}
\description{
Detect city by coordinates
}
\note{
The \code{find_city} function was added in version 0.3.0
}
\examples{
\dontrun{
find_city(longitude = 122.940625, latitude = 24.4520833334)
}
}
