% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city.R
\name{find_city}
\alias{find_city}
\title{Detect city by coordinates}
\usage{
find_city(longitude, latitude, ...)
}
\arguments{
\item{longitude}{longitude}

\item{latitude}{latitude}

\item{...}{export parameter to other functions}
}
\description{
Detect city by coordinates
}
\note{
The \code{find_city} function was added in version 0.3.0
}
\examples{
\dontrun{
find_city(longitude = 140.1137418, latitude = 36.0533957)
}
}
