% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city.R
\name{find_city}
\alias{find_city}
\title{Detect city by coordinates}
\usage{
find_city(longitude, latitude, geometry = NULL, ...)
}
\arguments{
\item{longitude}{longitude}

\item{latitude}{latitude}

\item{geometry}{XY sfg object}

\item{...}{export parameter to other functions}
}
\description{
Detect city by coordinates
}
\note{
The \code{find_city} function was added in version 0.3.0
}
\examples{
\dontrun{
find_city(longitude = 140.1137418, latitude = 36.0533957)

# Refrenced by sf geometry
library(sf)
find_city(geometry = st_point(c(136.6833, 35.05)))
}
}
