% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_name_table.R
\name{find_jis_code}
\alias{find_jis_code}
\title{Find prefecture code}
\usage{
find_jis_code(pref_code, admin_name, strict = TRUE)
}
\arguments{
\item{pref_code}{jis code from 1 to 47}

\item{admin_name}{prefecture names (string)}

\item{strict}{matching patterns}
}
\value{
Prefecture Identification Code (JIS X 0401)
}
\description{
Find prefecture code
}
\examples{
\dontrun{
find_jis_code(33, "\\u5009\\u6577\\u5e02")
find_jis_code(33, "\\u5009\\u6577\\u5e02", strict = FALSE)
}
}
