% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_name_table.R
\name{find_jis_code}
\alias{find_jis_code}
\title{Find JIS city code}
\usage{
find_jis_code(pref_code, admin_name, strict = TRUE)
}
\arguments{
\item{pref_code}{jis code from 1 to 47}

\item{admin_name}{prefecture names (string)}

\item{strict}{matching patterns}
}
\value{
Identification code for cities, towns and villages (JIS X 0402:2010)
}
\description{
Find JIS city code
}
