% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city.R
\name{find_pref}
\alias{find_pref}
\title{Detect prefecture by coordinates}
\usage{
find_pref(lon = lon, lat = lat)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}
}
\description{
Detect prefecture by coordinates
}
\note{
The \code{find_pref} function was added in version 0.2.1
}
