% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city.R
\name{find_pref}
\alias{find_pref}
\title{Detect prefecture by coordinates}
\usage{
find_pref(longitude, latitude, geometry = NULL, ...)
}
\arguments{
\item{longitude}{longitude}

\item{latitude}{latitude}

\item{geometry}{XY sfg object}

\item{...}{export parameter to other functions}
}
\description{
Detect prefecture by coordinates
}
\note{
The \code{find_pref} function was added in version 0.3.0
}
\examples{
\dontrun{
find_pref(longitude = 130.4412895, latitude = 30.2984335)

# Referenced by sf geometry
library(sf)
find_pref(geometry = st_point(c(130.4412895, 30.2984335)))
}
}
