% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city.R
\name{find_prefs}
\alias{find_prefs}
\title{Detect prefectures by coordinates}
\usage{
find_prefs(longitude, latitude)
}
\arguments{
\item{longitude}{longitude}

\item{latitude}{latitude}
}
\description{
Detect prefectures by coordinates
}
\examples{
\dontrun{
find_prefs(longitude = 122.940625, latitude = 24.4520833334)
find_prefs(longitude = 140.1137418, latitude = 36.0533957)
}
}
