% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jpn.R
\name{gg_jpn_district}
\alias{gg_jpn_district}
\title{Display city map based on 'ggplot2'}
\usage{
gg_jpn_district(df, ghost = TRUE, ...)
}
\arguments{
\item{df}{data frame}

\item{ghost}{logical (default TRUE). If don't spirit of ggplot2 calls, specify FALSE.}

\item{...}{Other arguments passed on to coord_map}
}
\description{
Rendering city map using 'ggplot2' and 'ggghost' systems.
}
\examples{
\dontrun{
gg_jpn_district(spdf_jpn_pref(code = 33, district = TRUE))
}
}

