% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jpnprefs}
\alias{jpnprefs}
\title{Prefectural informations in Japan}
\format{A data frame with 47 rows 7 variables:
\itemize{
  \item{jis_code: jis code}
  \item{prefecture: prefecture names}
  \item{capital: capital name for prefecture}
  \item{region: region}
  \item{major_island: }
  \item{capital_latitude: latitude for catital}
  \item{capital_longitude: longitude for catital}
}}
\usage{
jpnprefs
}
\description{
Prefectures dataset.
}
\keyword{datasets}

