% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mesh.R
\name{mesh_district}
\alias{mesh_district}
\title{Export district's mesh polygon}
\usage{
mesh_district(jis_code = NULL)
}
\arguments{
\item{jis_code}{jis code for prefecture and city identifical number.
If prefecture, must be from 1 to 47. If city, range of 5 digits.}
}
\description{
Export district's mesh polygon
}
\examples{
\dontrun{
mesh_district(jis_code = "33101")
mesh_distinct(jis_code = "05")
}
}
