% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prefecture_mesh}
\alias{prefecture_mesh}
\title{Prefecture's meshcode}
\format{A simple feature data frame with 314 rows 5 variables:
\itemize{
\item{prefcode: prefecture code}
\item{meshcode}
\item{name}
\item{type}
\item{geometry}
}}
\usage{
prefecture_mesh
}
\description{
Prefectures dataset.
}
\keyword{datasets}
