% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_ksj_p34}
\alias{read_ksj_p34}
\title{Intermediate function}
\usage{
read_ksj_p34(code = NULL, path = NULL)
}
\arguments{
\item{code}{prefecture code (JIS X 0402)}

\item{path}{path to N03 shapefile (if already exist)}
}
\description{
Intermediate function
}

