% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdf_jpn.R
\name{spdf_jpn_cities}
\alias{spdf_jpn_cities}
\title{Spatial Data frame for city area polygons}
\usage{
spdf_jpn_cities(jis_code_pref, jis_code_city = NULL, admin_name = NULL)
}
\arguments{
\item{jis_code_pref}{jis code from 1 to 47}

\item{jis_code_city}{jis code for city as jis_code_pref + identifier number}

\item{admin_name}{city name}
}
\description{
City area polygon data. When an administrative name (jis_code_city) or code (jis_code_city)
is specified as an argument, the target city data is extracted. If neither is given,
it becomes the data of the target prefecture.
}
\examples{
spdf_jpn_cities(jis_code_pref = 33, jis_code_city = 33103)
spdf_jpn_cities(jis_code_pref = 33, jis_code_city = c(33103, 33104, 33205))
}

