% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdf_jpn.R
\name{spdf_jpn_city}
\alias{spdf_jpn_city}
\title{Spatial Data frame city single}
\usage{
spdf_jpn_city(df, jis_code, admin_name = NULL)
}
\arguments{
\item{df}{data frame}

\item{jis_code}{jis code}

\item{admin_name}{administration name (full lenght)}
}
\description{
Spatial Data frame city single
}
\examples{
\dontrun{
spdf_jpn_pref(jis_code = 33, district = TRUE) \%>\%
  spdf_jpn_city(33103)
}
}

