% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdf_jpn.R
\name{spdf_jpn_pref}
\alias{spdf_jpn_pref}
\title{Spatial Data frame for prefecture area polygon}
\usage{
spdf_jpn_pref(code = NULL, admin_name = NULL, district = TRUE,
  download = FALSE)
}
\arguments{
\item{code}{jis code from 1 to 47}

\item{admin_name}{string}

\item{district}{logical (default TRUE)}

\item{download}{logical (default FALSE).}
}
\description{
Prefecture polygon data.
}
\details{
Collect unit of prefecture SpatialPolygonsDataFrame. If downalod argument is TRUE,
download administrative area data from the National Land Numeral Information Download Service (for law data).
}
\examples{
\dontrun{
spdf_jpn_pref(code = 33, district = FALSE)
}

}

