% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{which_pol_min}
\alias{which_pol_min}
\title{Internal function}
\usage{
which_pol_min(longitude, latitude, ...)
}
\arguments{
\item{longitude}{longitude}

\item{latitude}{latitude}

\item{...}{export parameter to other functions}
}
\description{
Internal function
}
